requirements.txt
github3.py
requests
tenacity
tqdm